move parsing and pieces to separate files
